py/gc: Speed up incremental GC cycles by tracking the high watermark.
In applications that use little memory and run GC regularly, the cost
of the sweep phase quickly becomes prohibitives as the amount of RAM
increases.

On an ESP32-S3 with 2 MB of external SPIRAM, for example, a trivial GC
cycle takes a minimum of 40ms, virtually all of it in the sweep phase.

Similarly, on the UNIX port with 1 GB of heap, a trivial GC takes 47 ms,
again virtually all of it in the sweep phase.

Speed up the sweep phase in the case most of the heap is empty by keeping
track of the ID of the highest block we allocated in an area since the
last GC.